requirements.txt
aiofiles==22.1.0
anyio==3.6.2
click==8.1.3
defusedxml==0.7.1
dnspython==2.2.1
fastapi==0.86.0
fonttools==4.38.0
fpdf2==2.5.7
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Pillow==9.3.0
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.2
PyPDF2==2.11.1
python-decouple==3.6
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.21.0
svg.path==6.2
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
